+++ src/Utility.hpp
@@ -5,7 +5,7 @@
 // Provides generic utility functions
 
 // Zeroes value if it's inside deadband range, and rescales values outside of it
-float ApplyDeadband(float value, float deadband);
+double ApplyDeadband(double value, double deadband);
 
 // Limits 'value' to within +- 'limit' (limit should be positive)
 template <class T>
@@ -22,4 +22,4 @@ T Limit(T value, T limit) {
 /* Rescales joystick value from [1..-1] to [0..rangeMax] (this includes flipping
  * the range)
  */
-float JoystickRescale(float value, float rangeMax);
+double JoystickRescale(double value, double rangeMax);
